moved logic
