+++ src/main/java/com/splitspendings/groupexpensesbackend/service/GroupInviteService.java
@@ -1,0 +1,17 @@
+package com.splitspendings.groupexpensesbackend.service;
+
+import com.splitspendings.groupexpensesbackend.dto.groupinvite.GroupInviteAcceptedDto;
+import com.splitspendings.groupexpensesbackend.dto.groupinvite.GroupInviteDto;
+import com.splitspendings.groupexpensesbackend.dto.groupinvite.NewGroupInviteDto;
+import com.splitspendings.groupexpensesbackend.model.GroupInvite;
+
+public interface GroupInviteService {
+
+    GroupInvite groupInviteModelById(Long inviteId);
+
+    GroupInviteDto createGroupInvite(NewGroupInviteDto newGroupInviteDto);
+
+    GroupInviteAcceptedDto acceptGroupInvite(Long inviteId);
+
+    void declineGroupInvite(Long id);
+}
