+++ src/main/java/com/splitspendings/groupexpensesbackend/service/GroupService.java
@@ -5,12 +5,8 @@ import com.splitspendings.groupexpensesbackend.dto.group.GroupInfoDto;
 import com.splitspendings.groupexpensesbackend.dto.group.GroupSpendingsDto;
 import com.splitspendings.groupexpensesbackend.dto.group.NewGroupDto;
 import com.splitspendings.groupexpensesbackend.dto.group.UpdateGroupInfoDto;
-import com.splitspendings.groupexpensesbackend.dto.groupinvite.GroupInviteAcceptedDto;
-import com.splitspendings.groupexpensesbackend.dto.groupinvite.GroupInviteDto;
-import com.splitspendings.groupexpensesbackend.dto.groupinvite.NewGroupInviteDto;
 import com.splitspendings.groupexpensesbackend.dto.groupmembership.GroupMembershipDto;
 import com.splitspendings.groupexpensesbackend.model.Group;
-import com.splitspendings.groupexpensesbackend.model.GroupInvite;
 
 import java.util.UUID;
 
@@ -27,14 +23,6 @@ public interface GroupService {
     GroupActiveMembersDto groupActiveMembersById(Long id);
 
     GroupMembershipDto groupMembership(Long id, UUID appUserId);
-
-    GroupInviteDto createGroupInvite(NewGroupInviteDto newGroupInviteDto);
-
-    GroupInvite groupInviteModelById(Long inviteId);
-
-    GroupInviteAcceptedDto acceptGroupInvite(Long inviteId);
-
-    void declineGroupInvite(Long id);
 
     void leaveGroup(Long id);
